requirements.txt

flask==2.3.3
python-telegram-bot==20.6
requests==2.31.0
gunicorn==21.2.0
pillow==10.0.0
python-dotenv==1.0.0
